+++ main.js
@@ -201,9 +201,13 @@ class HuumSauna extends utils.Adapter {
 
 			// @ts-ignore
 			if (steamerErrorstate && humstate && steamerErrorstate.val && humstate.val > 0) {
-				this.switchSauna(false);
-				this.setState("statusMessage", `Error:  and no water in steamer: -> ${this.huum.humidity * 10}%`, true);
+				this.setState("humidity",0,true);
 				this.log.warn(`Sauna switched off! Steam Mode with ${this.huum.humidity * 10}% and no water in steamer `);
+
+				//this.switchSauna(false);
+				this.setState("statusMessage", `Error: no water in steamer: ${this.huum.humidity * 10}%`, true);
+				this.log.warn(`Sauna humdity ${this.huum.humidity * 10}% set to 0% `);
+
 			}
 		}
 	}