+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 <!--
     ## **WORK IN PROGRESS**
 -->
+## **WORK IN PROGRESS**
+- Steamer Error -> setting humdidity to 0, old version otherwise sauna will be switched off
+
 ## 0.4.4 (2023-02-12)
 - adapted temp reached when sauna is stopped from the app
 - added an offset for temp reached value when use the intelligent sauna mode